backend/requirements.txt
# Core FastAPI and web server
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
starlette==0.27.0

# Database (MongoDB)
motor==3.7.1
pymongo==4.13.2

# Authentication and Security
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-multipart==0.0.6

# Image Storage (Cloudinary)
cloudinary==1.44.1
pillow==11.3.0

# Configuration and Environment
python-dotenv==1.1.1

# Validation and Data Processing
email-validator==2.1.0
